Add SNS resource
